move and adapt test class
