--- application/src/test/java/uk/gov/hmcts/ccd/definition/store/swagger/SwaggerPublisher.java
@@ -1,54 +1,0 @@
-package uk.gov.hmcts.ccd.definition.store.swagger;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.web.context.WebApplicationContext;
-
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-@SpringJUnitWebConfig
-@SpringBootTest
-@AutoConfigureMockMvc
-public class SwaggerPublisher {
-
-    private MockMvc mvc;
-
-    @Autowired
-    private WebApplicationContext webAppContext;
-
-    @BeforeEach
-    public void setup() {
-        this.mvc = webAppContextSetup(webAppContext).build();
-    }
-
-
-    @DisplayName("Generate swagger documentation")
-    @Test
-    @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
-    void generateDocs() throws Exception {
-        ResultActions perform = mvc.perform(get("/v2/api-docs"));
-        byte[] specs = perform
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsByteArray();
-
-        try (OutputStream outputStream = Files.newOutputStream(Paths.get("/tmp/swagger-specs.json"))) {
-            outputStream.write(specs);
-        }
-
-    }
-}
