+++ application/src/test/java/uk/gov/hmcts/net/ccd/definition/store/SwaggerPublisherTest.java
@@ -1,0 +1,32 @@
+package uk.gov.hmcts.net.ccd.definition.store;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.test.web.servlet.ResultActions;
+
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
+public class SwaggerPublisherTest extends BaseTest {
+
+    @DisplayName("Generate swagger documentation")
+    @Test
+    @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
+    void generateDocs() throws Exception {
+        ResultActions perform = mockMvc.perform(get("/v2/api-docs"));
+        byte[] specs = perform
+            .andExpect(status().isOk())
+            .andReturn()
+            .getResponse()
+            .getContentAsByteArray();
+
+        try (OutputStream outputStream = Files.newOutputStream(Paths.get("/tmp/swagger-specs.json"))) {
+            outputStream.write(specs);
+        }
+
+    }
+}
